requirements.txt
Flask==2.3.3
python-telegram-bot[asyncio]==20.6
google-cloud-dialogflow==2.24.0
python-dotenv==1.0.0